AB#87720 CR fixes.
