+++ src/HE.Investment.AHP.Domain/Delivery/CommandHandlers/DeliveryCommandHandlerBase.cs
@@ -1,6 +1,13 @@
+using System.Diagnostics.CodeAnalysis;
+using HE.Investment.AHP.Domain.Delivery.Commands;
+using HE.Investment.AHP.Domain.Delivery.Entities;
+using HE.Investment.AHP.Domain.Delivery.Repositories;
+using HE.Investments.Account.Shared;
 using HE.Investments.Common.Exceptions;
 using HE.Investments.Common.Validators;
+using MediatR;
 using Microsoft.Extensions.Logging;
+using ApplicationId = HE.Investment.AHP.Domain.Application.ValueObjects.ApplicationId;
 
 namespace HE.Investment.AHP.Domain.Delivery.CommandHandlers;
 
@@ -31,4 +38,36 @@ public abstract class DeliveryCommandHandlerBase
 
         return errors;
     }
+}
+
+[SuppressMessage("Maintainability Rules", "SA1402", Justification = "Generic version of the same command handler.")]
+public abstract class DeliveryCommandHandlerBase<TCommand> : DeliveryCommandHandlerBase, IRequestHandler<TCommand, OperationResult>
+    where TCommand : DeliveryCommandBase
+{
+    private readonly IDeliveryPhaseRepository _repository;
+
+    private readonly IAccountUserContext _accountUserContext;
+
+    protected DeliveryCommandHandlerBase(IDeliveryPhaseRepository repository, IAccountUserContext accountUserContext, ILogger logger)
+        : base(logger)
+    {
+        _accountUserContext = accountUserContext;
+        _repository = repository;
+    }
+
+    public async Task<OperationResult> Handle(TCommand request, CancellationToken cancellationToken)
+    {
+        var account = await _accountUserContext.GetSelectedAccount();
+        var deliveryPhases = await _repository.GetByApplicationId(new ApplicationId(request.ApplicationId), account, cancellationToken);
+        var validationErrors = Perform(deliveryPhases, request);
+        if (validationErrors.Any())
+        {
+            return new OperationResult(validationErrors);
+        }
+
+        await _repository.Save(deliveryPhases, account.SelectedOrganisationId(), cancellationToken);
+        return OperationResult.Success();
+    }
+
+    protected abstract IList<ErrorItem> Perform(DeliveryPhasesEntity deliveryPhases, TCommand request);
 }
