+++ src/HE.Investment.AHP.Domain/Delivery/CommandHandlers/RemoveDeliveryPhaseCommandHandler.cs
@@ -1,42 +1,25 @@
 using HE.Investment.AHP.Domain.Delivery.Commands;
+using HE.Investment.AHP.Domain.Delivery.Entities;
 using HE.Investment.AHP.Domain.Delivery.Repositories;
 using HE.Investment.AHP.Domain.Delivery.ValueObjects;
 using HE.Investments.Account.Shared;
 using HE.Investments.Common.Validators;
-using MediatR;
 using Microsoft.Extensions.Logging;
-using ApplicationId = HE.Investment.AHP.Domain.Application.ValueObjects.ApplicationId;
 
 namespace HE.Investment.AHP.Domain.Delivery.CommandHandlers;
 
-public class RemoveDeliveryPhaseCommandHandler : DeliveryCommandHandlerBase, IRequestHandler<RemoveDeliveryPhaseCommand, OperationResult>
+public class RemoveDeliveryPhaseCommandHandler : DeliveryCommandHandlerBase<RemoveDeliveryPhaseCommand>
 {
-    private readonly IDeliveryPhaseRepository _repository;
-
-    private readonly IAccountUserContext _accountUserContext;
-
     public RemoveDeliveryPhaseCommandHandler(
         IDeliveryPhaseRepository repository,
         IAccountUserContext accountUserContext,
         ILogger<RemoveDeliveryPhaseCommandHandler> logger)
-        : base(logger)
+        : base(repository, accountUserContext, logger)
     {
-        _repository = repository;
-        _accountUserContext = accountUserContext;
     }
 
-    public async Task<OperationResult> Handle(RemoveDeliveryPhaseCommand request, CancellationToken cancellationToken)
+    protected override IList<ErrorItem> Perform(DeliveryPhasesEntity deliveryPhases, RemoveDeliveryPhaseCommand request)
     {
-        var account = await _accountUserContext.GetSelectedAccount();
-        var deliveryPhases = await _repository.GetByApplicationId(new ApplicationId(request.ApplicationId), account, cancellationToken);
-        var validationErrors = PerformWithValidation(
-            () => deliveryPhases.Remove(new DeliveryPhaseId(request.DeliveryPhaseId), request.RemoveDeliveryPhaseAnswer));
-        if (validationErrors.Any())
-        {
-            return new OperationResult(validationErrors);
-        }
-
-        await _repository.Save(deliveryPhases, account.SelectedOrganisationId(), cancellationToken);
-        return OperationResult.Success();
+        return PerformWithValidation(() => deliveryPhases.Remove(new DeliveryPhaseId(request.DeliveryPhaseId), request.RemoveDeliveryPhaseAnswer));
     }
 }
