+++ src/HE.Investment.AHP.Domain/Delivery/Commands/DeliveryCommandBase.cs
@@ -1,0 +1,6 @@
+using HE.Investments.Common.Validators;
+using MediatR;
+
+namespace HE.Investment.AHP.Domain.Delivery.Commands;
+
+public abstract record DeliveryCommandBase(string ApplicationId) : IRequest<OperationResult>;
