+++ src/HE.Investment.AHP.Domain/Delivery/Commands/RemoveDeliveryPhaseCommand.cs
@@ -1,6 +1,4 @@
 using HE.Investment.AHP.Contract.Delivery.Enums;
-using HE.Investments.Common.Validators;
-using MediatR;
 
 namespace HE.Investment.AHP.Domain.Delivery.Commands;
 
@@ -8,4 +6,4 @@ public record RemoveDeliveryPhaseCommand(
         string ApplicationId,
         string DeliveryPhaseId,
         RemoveDeliveryPhaseAnswer RemoveDeliveryPhaseAnswer)
-    : IRequest<OperationResult>;
+    : DeliveryCommandBase(ApplicationId);
